requirements.txt
azure-core==1.27.0
azure-identity==1.13.0
azure-storage-blob==12.16.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Flask==2.3.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.22.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
pywin32==306
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.4
